hw/omap3_mmc.c: Set TC status bit for R1b no-data commands
We emulate MMC commands with response type R1b and no data transfer
instantaneously, so we must set the TC bit as well as the CC bit,
to indicate that the "busy" period has ended.

Without this fix the Linux driver will hang waiting for a TC interrupt
that never happens when it tries to do a block erase.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>